fix: Type error: Route "app/api/auth/[...nextauth]/route.ts" during deploy
