+++ src/app/add-product/page.tsx
@@ -5,7 +5,8 @@ import { getAvailableSizes } from "@/lib/functions/getAvailableSizes";
 import { Metadata } from "next";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import authOptions from "../api/auth/[...nextauth]/options";
+// import { authOptions } from "../api/auth/[...nextauth]/route";
 
 export const metadata: Metadata = {
   title: "Add product",
