+++ src/app/api/auth/[...nextauth]/options.ts
@@ -1,0 +1,31 @@
+import { prisma } from "@/lib/db/prisma";
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import { NextAuthOptions } from "next-auth";
+import { Adapter } from "next-auth/adapters";
+import GoogleProvider from "next-auth/providers/google";
+import { env } from "@/lib/env";
+import mergeAnonymousCartIntoUserCart from "@/lib/db/cart";
+import { PrismaClient } from "@prisma/client";
+
+const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma as PrismaClient) as Adapter,
+  providers: [
+    GoogleProvider({
+      clientId: env.GOOGLE_CLIENT_ID,
+      clientSecret: env.GOOGLE_CLIENT_SECRET,
+    }),
+  ],
+  callbacks: {
+    session({ session, user }) {
+      session.user.id = user.id;
+      return session;
+    },
+  },
+  events: {
+    async signIn({ user }) {
+      await mergeAnonymousCartIntoUserCart(user.id);
+    },
+  },
+};
+
+export default authOptions;
