+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,32 +1,5 @@
-import { prisma } from "@/lib/db/prisma";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { NextAuthOptions } from "next-auth";
-import { Adapter } from "next-auth/adapters";
-import GoogleProvider from "next-auth/providers/google";
-import NextAuth from "next-auth/next";
-import { env } from "@/lib/env";
-import mergeAnonymousCartIntoUserCart from "@/lib/db/cart";
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma) as Adapter,
-  providers: [
-    GoogleProvider({
-      clientId: env.GOOGLE_CLIENT_ID,
-      clientSecret: env.GOOGLE_CLIENT_SECRET,
-    }),
-  ],
-  callbacks: {
-    session({ session, user }) {
-      session.user.id = user.id;
-      return session;
-    },
-  },
-  events: {
-    async signIn({ user }) {
-      await mergeAnonymousCartIntoUserCart(user.id);
-    },
-  },
-};
+import NextAuth from "next-auth";
+import authOptions from "./options";
 
 const handler = NextAuth(authOptions);
